agente_testes/app/requirements.txt
fastapi
uvicorn[standard]
langgraph
openai
python-dotenv
langchain 
langchain-openai
langgraph-checkpoint-sqlite
pydantic[email]
asyncpg